+++ CHANGELOG.md
@@ -7,6 +7,13 @@ All changes to the Sawtooth Water Grant project are recorded here. Changes must 
 
 Thanks for contributing.
 
+## [0.4]
+
+### Added
+
+- Admin Transaction Processor
+- Admin Subscriber
+
 ## [0.3]
 
 ### Updated